refactor engine with scan and pipeline
