+++ internal/engine/engine.go
@@ -1,313 +1,53 @@
-// internal/engine/engine.go
 package engine
 
 import (
 	"bytes"
 	"context"
-	"errors"
 	"fmt"
 	"io"
-	"log"
 	"os"
-	"path/filepath"
-	"sort"
-	"strings"
-	"sync"
-	"sync/atomic"
 
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclwrite"
 
 	"github.com/oferchen/hclalign/config"
+	"github.com/oferchen/hclalign/internal/align"
 	"github.com/oferchen/hclalign/internal/diff"
 	internalfs "github.com/oferchen/hclalign/internal/fs"
-	"github.com/oferchen/hclalign/internal/hclalign"
-	"github.com/oferchen/hclalign/patternmatching"
 )
 
 var (
 	testHookAfterParse   func()
 	testHookAfterReorder func()
-	reorderAttributes    = hclalign.ReorderAttributes
 	WriteFileAtomic      = internalfs.WriteFileAtomic
 )
 
-func Process(ctx context.Context, cfg *config.Config) (bool, error) {
+// Run is the entrypoint for processing according to cfg. It delegates to
+// ProcessReader when reading from STDIN, otherwise it scans targets and runs the
+// processing pipeline.
+func Run(ctx context.Context, cfg *config.Config) (bool, error) {
 	if cfg.Stdin {
 		return processReader(ctx, os.Stdin, os.Stdout, cfg)
 	}
-	return processFiles(ctx, cfg)
-}
-
-func processFiles(ctx context.Context, cfg *config.Config) (bool, error) {
-	if _, err := os.Stat(cfg.Target); err != nil {
-		if os.IsNotExist(err) {
-			return false, fmt.Errorf("target %q does not exist", cfg.Target)
-		}
-		return false, err
-	}
-	matcher, err := patternmatching.NewMatcher(cfg.Include, cfg.Exclude, cfg.Target)
+	files, err := Scan(ctx, cfg)
 	if err != nil {
 		return false, err
 	}
-	var files []string
-	var walk func(context.Context, string) error
-	walk = func(ctx context.Context, dir string) error {
-		if !matcher.Matches(dir) {
-			return nil
-		}
-		if err := ctx.Err(); err != nil {
-			return err
-		}
-		entries, err := os.ReadDir(dir)
-		if err != nil {
-			return err
-		}
-		for _, entry := range entries {
-			if err := ctx.Err(); err != nil {
-				return err
-			}
-			path := filepath.Join(dir, entry.Name())
-			if entry.Type()&os.ModeSymlink != 0 {
-				info, err := os.Stat(path)
-				if err != nil {
-					return err
-				}
-				if !cfg.FollowSymlinks {
-					continue
-				}
-				if info.IsDir() {
-					if err := walk(ctx, path); err != nil {
-						return err
-					}
-					continue
-				}
-			}
-			if entry.IsDir() {
-				if err := walk(ctx, path); err != nil {
-					return err
-				}
-				continue
-			}
-			if matcher.Matches(path) {
-				files = append(files, path)
-			}
-		}
-		return nil
-	}
-	info, err := os.Lstat(cfg.Target)
+	changed, outs, err := Pipeline(ctx, files, cfg)
 	if err != nil {
-		return false, err
+		return changed, err
 	}
-	if info.Mode()&os.ModeSymlink != 0 {
-		resolved, err := os.Stat(cfg.Target)
-		if err != nil {
-			return false, err
-		}
-		if resolved.IsDir() {
-			if cfg.FollowSymlinks {
-				if err := walk(ctx, cfg.Target); err != nil {
-					return false, err
-				}
-			}
-		} else if cfg.FollowSymlinks {
-			if matcher.Matches(cfg.Target) {
-				files = append(files, cfg.Target)
-			}
-		}
-	} else if info.IsDir() {
-		if err := walk(ctx, cfg.Target); err != nil {
-			return false, err
-		}
-	} else {
-		if matcher.Matches(cfg.Target) {
-			files = append(files, cfg.Target)
-		}
-	}
-	sort.Strings(files)
-
-	type result struct {
-		path string
-		data []byte
-	}
-
-	var changed atomic.Bool
-	outs := make(map[string][]byte, len(files))
-
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	fileCh := make(chan string)
-	results := make(chan result, len(files))
-
-	var wg sync.WaitGroup
-	var errMu sync.Mutex
-	var errs []error
-
-	go func() {
-		defer close(fileCh)
-		for _, f := range files {
-			if ctx.Err() != nil {
-				return
-			}
-			select {
-			case fileCh <- f:
-			case <-ctx.Done():
-				return
-			}
-		}
-	}()
-
-	for i := 0; i < cfg.Concurrency; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			for {
-				select {
-				case <-ctx.Done():
-					return
-				case f, ok := <-fileCh:
-					if !ok {
-						return
-					}
-					ch, out, err := processSingleFile(ctx, f, cfg)
-					if err != nil {
-						if !errors.Is(err, context.Canceled) {
-							log.Printf("error processing file %s: %v", f, err)
-							errMu.Lock()
-							errs = append(errs, fmt.Errorf("%s: %w", f, err))
-							errMu.Unlock()
-							cancel()
-						}
-						return
-					}
-					if ch {
-						changed.Store(true)
-					}
-					if len(out) > 0 {
-						select {
-						case results <- result{path: f, data: out}:
-						case <-ctx.Done():
-							return
-						}
-					}
-					if cfg.Verbose {
-						log.Printf("processed file: %s", f)
-					}
-				}
-			}
-		}()
-	}
-
-	wg.Wait()
-	close(results)
-	for r := range results {
-		outs[r.path] = r.data
-	}
-
 	for _, f := range files {
 		if out, ok := outs[f]; ok && len(out) > 0 {
-			if err := ctx.Err(); err != nil {
-				if !errors.Is(err, context.Canceled) {
-					return changed.Load(), err
-				}
-				break
-			}
 			if _, err := fmt.Fprintf(os.Stdout, "\n--- %s ---\n", f); err != nil {
-				return changed.Load(), err
+				return changed, err
 			}
 			if _, err := os.Stdout.Write(out); err != nil {
-				return changed.Load(), err
+				return changed, err
 			}
 		}
 	}
-
-	if len(errs) > 0 {
-		messages := make([]string, len(errs))
-		for i, err := range errs {
-			messages[i] = err.Error()
-		}
-		return changed.Load(), errors.New(strings.Join(messages, "\n"))
-	}
-	if err := ctx.Err(); err != nil {
-		return changed.Load(), err
-	}
-
-	return changed.Load(), nil
-}
-
-func processSingleFile(ctx context.Context, filePath string, cfg *config.Config) (bool, []byte, error) {
-	if err := ctx.Err(); err != nil {
-		return false, nil, err
-	}
-	data, perm, hints, err := internalfs.ReadFileWithHints(ctx, filePath)
-	if err != nil {
-		return false, nil, fmt.Errorf("error reading file %s: %w", filePath, err)
-	}
-	if err := ctx.Err(); err != nil {
-		return false, nil, err
-	}
-
-	file, diags := hclwrite.ParseConfig(data, filePath, hcl.InitialPos)
-	if diags.HasErrors() {
-		return false, nil, fmt.Errorf("parsing error in file %s: %v", filePath, diags.Errs())
-	}
-	if testHookAfterParse != nil {
-		testHookAfterParse()
-	}
-	if err := ctx.Err(); err != nil {
-		return false, nil, err
-	}
-
-	if err := reorderAttributes(file, cfg.Order, cfg.StrictOrder); err != nil {
-		return false, nil, err
-	}
-	if testHookAfterReorder != nil {
-		testHookAfterReorder()
-	}
-	if err := ctx.Err(); err != nil {
-		return false, nil, err
-	}
-
-	formatted := bytes.ReplaceAll(file.Bytes(), []byte("\r\n"), []byte("\n"))
-	styled := internalfs.ApplyHints(formatted, hints)
-	original := data
-	if bom := hints.BOM(); len(bom) > 0 {
-		original = append(append([]byte{}, bom...), original...)
-	}
-	changed := !bytes.Equal(original, styled)
-
-	var out []byte
-
-	switch cfg.Mode {
-	case config.ModeWrite:
-		if !changed {
-			if cfg.Stdout {
-				out = styled
-			}
-			return false, out, nil
-		}
-		if err := WriteFileAtomic(ctx, filePath, formatted, perm, hints); err != nil {
-			return false, nil, fmt.Errorf("error writing file %s with original permissions: %w", filePath, err)
-		}
-		if cfg.Stdout {
-			out = styled
-		}
-	case config.ModeCheck:
-		if cfg.Stdout {
-			out = styled
-		}
-	case config.ModeDiff:
-		if changed {
-			text, err := diff.Unified(filePath, filePath, original, styled, hints.Newline)
-			if err != nil {
-				return false, nil, err
-			}
-			out = []byte(text)
-		}
-	}
-
-	return changed, out, nil
+	return changed, nil
 }
 
 func processReader(ctx context.Context, r io.Reader, w io.Writer, cfg *config.Config) (bool, error) {
@@ -325,14 +65,17 @@ func processReader(ctx context.Context, r io.Reader, w io.Writer, cfg *config.Co
 		data = data[len(hints.BOM()):]
 	}
 
-	file, diags := hclwrite.ParseConfig(data, "stdin", hcl.InitialPos)
+	// Phase A: format the input prior to alignment.
+	formatted := hclwrite.Format(data)
+	file, diags := hclwrite.ParseConfig(formatted, "stdin", hcl.InitialPos)
 	if diags.HasErrors() {
 		return false, fmt.Errorf("parsing error: %v", diags.Errs())
 	}
-	if err := hclalign.ReorderAttributes(file, cfg.Order, cfg.StrictOrder); err != nil {
+	if err := align.Apply(file, cfg); err != nil {
 		return false, err
 	}
-	formatted := bytes.ReplaceAll(file.Bytes(), []byte("\r\n"), []byte("\n"))
+
+	formatted = bytes.ReplaceAll(file.Bytes(), []byte("\r\n"), []byte("\n"))
 	styled := internalfs.ApplyHints(formatted, hints)
 	original := data
 	if bom := hints.BOM(); len(bom) > 0 {
