+++ internal/engine/process_reader_test.go
@@ -97,11 +97,11 @@ func TestProcessPrintsDelimiters(t *testing.T) {
 	require.NoError(t, os.WriteFile(f2, []byte("variable \"b\" {}\n"), 0o644))
 
 	cfg := &config.Config{
-		Target:		dir,
-		Include:	[]string{"**/*.tf"},
-		Mode:		config.ModeCheck,
-		Stdout:		true,
-		Concurrency:	1,
+		Target:      dir,
+		Include:     []string{"**/*.tf"},
+		Mode:        config.ModeCheck,
+		Stdout:      true,
+		Concurrency: 1,
 	}
 
 	r, w, err := os.Pipe()
@@ -110,7 +110,7 @@ func TestProcessPrintsDelimiters(t *testing.T) {
 	os.Stdout = w
 	t.Cleanup(func() { os.Stdout = stdout })
 
-	_, err = engine.Process(context.Background(), cfg)
+	_, err = engine.Run(context.Background(), cfg)
 	require.NoError(t, err)
 	w.Close()
 	out, err := io.ReadAll(r)
@@ -120,4 +120,3 @@ func TestProcessPrintsDelimiters(t *testing.T) {
 	require.Contains(t, s, fmt.Sprintf("\n--- %s ---\n", f1))
 	require.Contains(t, s, fmt.Sprintf("\n--- %s ---\n", f2))
 }
-
