+++ tests/cases/inline_comment_after_brace/out.tf
@@ -1,7 +1,5 @@
 // example8.tf
 variable "example_var_inline_comment" { # Problematic corner case inline comment
-  /* multi line
-  comment */
   description = "An example variable with an inline comment" # Just another inline
   // different tyoe of comment
   type        = string
